Refactor bmv2 handlers into handlers/ package by domain
- Move Driver and all HTTP handlers out of api/handlers.go into a dedicated handlers/ package, split by function.
- Extract shared writeJSONError helper and downloadFile/compileP4 utilities.
- Update Dockerfile to copy the new directory.
