+++ targets/bmv2/Dockerfile
@@ -14,6 +14,7 @@ RUN go mod download
 # Copy the go source
 COPY cmd/main.go cmd/main.go
 COPY api/ api/
+COPY handlers/ handlers/
 
 # Build
 # the GOARCH has not a default value to allow the binary be built according to the host where the command
